refactor: HttpRequest を自動で行う部分の関数を切り出し
