+++ pages/wait.tsx
@@ -11,47 +11,16 @@ import { PageBackIcon } from "@/components/common/PageBackIcon";
 import { VSpacer } from "@/components/common/Spacer";
 import { MemberList } from "@/components/MemberList";
 
-import {
-  BASE_URL,
-  INTERVAL,
-  IS_AUTO_REQUEST,
-  MAX_AUTO_HTTP_REQUEST,
-} from "@/data/data";
+import { BASE_URL, IS_AUTO_REQUEST } from "@/data/data";
 
 import { RecoilOwner, RecoilRoom } from "@/store/Recoil";
 
 import { HandleError } from "@/hooks/useError";
+import { AutoHttpRequest } from "@/hooks/useHttpRequest";
 
 type Player = {
   nickname: string;
   particIcon: number;
-};
-
-let count = 0;
-let lastTime = Date.now();
-const AutoHTTPRequest = (Fun: () => void) => {
-  {
-    try {
-      const nowTime = Date.now();
-      if (count === 0 || lastTime + INTERVAL - 100 < nowTime) {
-        Fun();
-
-        lastTime = nowTime;
-        count += 1;
-
-        if (MAX_AUTO_HTTP_REQUEST < count) {
-          throw count;
-        }
-
-        setTimeout(() => {
-          AutoHTTPRequest(Fun);
-        }, INTERVAL);
-      }
-    } catch (count) {
-      // eslint-disable-next-line no-console
-      console.log(`Leave AutoHTTPRequest Loop`);
-    }
-  }
 };
 
 const Wait: NextPage = () => {
@@ -96,9 +65,6 @@ const Wait: NextPage = () => {
   const handleUpdate = () => {
     judgeIsGameStart();
     updateParticList();
-    console.log("####################");
-    console.log("#####  REQUEST  ####");
-    console.log("####################");
   };
 
   const handleGameStart = () => {
@@ -119,7 +85,7 @@ const Wait: NextPage = () => {
   useEffect(() => {
     if (IS_AUTO_REQUEST) {
       // 最初に1回 AutoHTTPRequest 関数を実行すれば、その後は再帰する
-      AutoHTTPRequest(handleUpdate);
+      AutoHttpRequest(handleUpdate, 0, Date.now());
     } else {
       handleUpdate();
     }
